# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Burgertown",
    "version": "1.1.0",
    "description": "HTTP APIs for Burgertown on Oak Street, a fictional smash-burger restaurant.\nImport this document into Atlas (or any workflow generator). Burgertown is the dummy merchant; Atlas owns blast radius and workflow generation.\n\nMoney is integer cents. Tax is 8.5% of (subtotal − discount − reward_discount). Delivery fee is not taxed.\nPOST /v1/payments requires amount_cents equal to check.totals.due_cents. Redeeming an offer changes due — re-fetch the check or you get 422 amount_mismatch.\nThe only retryable error is 503 no_courier_available (x-retryable on that response).\nProcessor payment_method: pm_ok | pm_decline (402) | pm_timeout (504).\nReset fixtures with POST /v1/sandbox { \"reset\": true } before scripted flows.\n\nSeeded happy-path ids: loc_oak, tbl_4 / chk_ok (due 1736), gst_maya (1240 gold), emp_jon, itm_townie, pm_ok, off_free_fries, dlv_pending, crr_sam.",
    "contact": {
      "name": "Burgertown",
      "email": "hello@burgertown.dev"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:43123",
      "description": "Burgertown Oak Street sandbox"
    }
  ],
...
  "paths": {
...
    "/v1/locations/{location_id}": {
      "get": {
        "operationId": "getLocation",
        "tags": [
          "Locations"
        ],
        "summary": "Get a location",
        "description": "Fetch one store by id.",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "location_not_found: Unknown location_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "location_not_found",
                    "message": "Unknown location_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "location_not_found",
                "when": "Unknown location_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "locations",
        "x-depends-on": []
      }
    },
    "/v1/locations/{location_id}/tables": {
      "get": {
        "operationId": "listTables",
        "tags": [
          "Floor"
        ],
        "summary": "List tables",
        "description": "Floor plan tables for a location.",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Table"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "location_not_found: Unknown location_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "location_not_found",
                    "message": "Unknown location_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "location_not_found",
                "when": "Unknown location_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "floor",
        "x-depends-on": [
          "locations",
          "fulfillment"
        ]
      }
    },
    "/v1/tables/{table_id}": {
      "get": {
        "operationId": "getTable",
        "tags": [
          "Floor"
        ],
        "summary": "Get a table",
        "description": "Includes current check_id when the table is occupied.",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Table"
                },
                "example": {
                  "id": "tbl_4",
                  "check_id": "chk_ok",
                  "status": "occupied"
                }
              }
            }
          },
          "404": {
            "description": "table_not_found: Unknown table_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "table_not_found",
                    "message": "Unknown table_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "table_not_found",
                "when": "Unknown table_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "floor",
        "x-depends-on": [
          "locations",
          "fulfillment"
        ]
      }
    },
    "/v1/locations/{location_id}/zones": {
      "get": {
        "operationId": "listZones",
        "tags": [
          "Zones"
        ],
        "summary": "List delivery zones",
        "description": "Zones with fee_cents and eta_minutes.",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Zone"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "location_not_found: Unknown location_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "location_not_found",
                    "message": "Unknown location_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "location_not_found",
                "when": "Unknown location_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "zones",
        "x-depends-on": [
          "locations"
        ]
      }
    },
    "/v1/zones/check": {
      "post": {
        "operationId": "checkZone",
        "tags": [
          "Zones"
        ],
        "summary": "Check an address against zones",
        "description": "Body: location_id, address_id. Hillsboro (addr_far) is out of zone.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneCheck"
                }
              }
            }
          },
          "404": {
            "description": "address_not_found: Unknown address_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "address_not_found",
                    "message": "Unknown address_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "address_not_found",
                "when": "Unknown address_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "zones",
        "x-depends-on": [
          "locations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "location_id",
                  "address_id"
                ],
                "properties": {
                  "location_id": {
                    "type": "string",
                    "examples": [
                      "loc_oak"
                    ]
                  },
                  "address_id": {
                    "type": "string",
                    "examples": [
                      "addr_maya"
                    ]
                  }
                }
              },
              "example": {
                "location_id": "loc_oak",
                "address_id": "addr_maya"
              }
            }
          }
        }
      }
    },
    "/v1/menus": {
      "get": {
        "operationId": "listMenus",
        "tags": [
          "Menus"
        ],
        "summary": "List menus",
        "description": "Menus for a location. Pass location_id as a query param.",
        "parameters": [
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "description": "Filter by location.",
            "schema": {
              "type": "string"
            },
            "example": "loc_oak"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Menu"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-burgertown-api": "menus",
        "x-depends-on": [
          "locations"
        ]
      }
    },
    "/v1/menus/{menu_id}": {
      "get": {
        "operationId": "getMenu",
        "tags": [
          "Menus"
        ],
        "summary": "Get a menu with categories",
        "description": "Full menu tree used before catalog item hydration.",
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Menu"
                }
              }
            }
          },
          "404": {
            "description": "menu_not_found: Unknown menu_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "menu_not_found",
                    "message": "Unknown menu_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "menu_not_found",
                "when": "Unknown menu_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "menus",
        "x-depends-on": [
          "locations"
        ]
      }
    },
    "/v1/catalog/items": {
      "get": {
        "operationId": "listItems",
        "tags": [
          "Catalog"
        ],
        "summary": "List catalog items",
        "description": "Optional exclude_allergens=alg_dairy hides dairy-tagged items.",
        "parameters": [
          {
            "name": "exclude_allergens",
            "in": "query",
            "required": false,
            "description": "Comma-separated allergen ids to hide. Example: alg_dairy.",
            "schema": {
              "type": "string"
            },
            "example": "alg_dairy"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CatalogItem"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-burgertown-api": "catalog",
        "x-depends-on": [
          "menus"
        ]
      }
    },
    "/v1/catalog/items/{item_id}": {
      "get": {
        "operationId": "getItem",
        "tags": [
          "Catalog"
        ],
        "summary": "Get a catalog item",
        "description": "Single item including price, tax_ids, and allergen_ids.",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "404": {
            "description": "item_not_found: Unknown item_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "item_not_found",
                    "message": "Unknown item_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "item_not_found",
                "when": "Unknown item_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "catalog",
        "x-depends-on": [
          "menus"
        ]
      },
      "patch": {
        "operationId": "patchItem",
        "tags": [
          "Catalog"
        ],
        "summary": "Update an item",
        "description": "Price edits flow into open checks on the next order send.",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "404": {
            "description": "item_not_found: Unknown item_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "item_not_found",
                    "message": "Unknown item_id",
                    "details": {}
                  }
                }
              }
            },
            "x-error-codes": [
              {
                "code": "item_not_found",
                "when": "Unknown item_id",
                "retryable": false
              }
            ]
          }
        },
        "x-burgertown-api": "catalog",
        "x-depends-on": [
          "menus"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "price": {
                    "type": "integer",
                    "description": "Integer cents. 1100 = $11.00.",
                    "examples": [
                      1100
                    ]
                  },
                  "name": {
                    "type": "string"
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/reports/sales": {
      "get": {
        "operationId": "salesReport",
        "tags": [
          "Reports"
        ],
        "summary": "Sales report",
        "description": "Net sales, tips, refunds, check counts. Optional group_by=fulfillment_type.",
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "description": "Pass fulfillment_type to split dine-in / pickup / delivery.",
            "schema": {
              "type": "string"
            },
            "example": "fulfillment_type"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesReport"
                }
              }
            }
          }
        },
        "x-burgertown-api": "reports",
        "x-depends-on": [
          "settlements",
          "checks",
          "payments",
          "delivery"
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "SandboxKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional. Any string is accepted in this sandbox."
      }
    }
  },
  "security": [
    {
      "SandboxKey": []
    }
  ]
}